+++ src/main/java/net/silentchaos512/gems/GemsConfig.java
@@ -1,38 +1,32 @@
 package net.silentchaos512.gems;
 
-import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.fml.event.config.ModConfigEvent;
 import net.neoforged.neoforge.common.ModConfigSpec;
 
-@EventBusSubscriber(modid = SilentGems.MOD_ID, bus = EventBusSubscriber.Bus.MOD)
 public final class GemsConfig {
+    public static final Common COMMON;
+    public static final ModConfigSpec COMMON_SPEC;
+
+    static {
+        var commonPair = new ModConfigSpec.Builder().configure(Common::new);
+        COMMON = commonPair.getLeft();
+        COMMON_SPEC = commonPair.getRight();
+    }
+
     public static final class Common {
-        private static final ModConfigSpec.Builder BUILDER = new ModConfigSpec.Builder();
+        public final ModConfigSpec.IntValue glowroseNormalLight;
+        public final ModConfigSpec.IntValue glowrosePottedLight;
 
-        // block
-        public static final ModConfigSpec.IntValue GLOWROSE_NORMAL_LIGHT = BUILDER
-                .comment("The light level of free-standing glowroses.",
-                        "Existing glowroses may not update until broken and replaced.")
-                .defineInRange("glowrose.normalLight", 10, 0, 15);
-        public static final ModConfigSpec.IntValue GLOWROSE_POTTED_LIGHT = BUILDER
-                .comment("The light level of glowroses planted in vanilla flower pots",
-                        "Existing blocks may not update until broken and replaced.")
-                .defineInRange("glowrose.pottedLight", 15, 0, 15);
-
-        static final ModConfigSpec SPEC = BUILDER.build();
-
-        public static int glowroseNormalLight;
-        public static int glowrosePottedLight;
-
-        private Common() {}
+        private Common(ModConfigSpec.Builder builder) {
+            glowroseNormalLight = builder
+                    .comment("The light level of free-standing glowroses.",
+                            "Existing glowroses may not update until broken and replaced.")
+                    .defineInRange("glowrose.normalLight", 10, 0, 15);
+            glowrosePottedLight = builder
+                    .comment("The light level of glowroses planted in vanilla flower pots.",
+                            "Existing blocks may not update until broken and replaced.")
+                    .defineInRange("glowrose.pottedLight", 15, 0, 15);
+        }
     }
 
     private GemsConfig() {}
-
-    @SubscribeEvent
-    public static void onLoad(final ModConfigEvent event) {
-        Common.glowroseNormalLight = Common.GLOWROSE_NORMAL_LIGHT.get();
-        Common.glowrosePottedLight = Common.GLOWROSE_POTTED_LIGHT.get();
-    }
 }