+++ src/main/java/net/silentchaos512/gems/SilentGems.java
@@ -2,12 +2,14 @@ package net.silentchaos512.gems;
 
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.RandomSource;
-import net.minecraft.world.item.CreativeModeTab;
+import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.ModList;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.config.ModConfig;
+import net.neoforged.fml.loading.FMLLoader;
+import net.silentchaos512.gems.client.SilentGemsClient;
 import net.silentchaos512.gems.setup.Registration;
 import net.silentchaos512.gems.util.TextUtil;
 import org.apache.logging.log4j.LogManager;
@@ -25,12 +27,14 @@ public class SilentGems {
     public static final RandomSource RANDOM_SOURCE = RandomSource.create();
     public static final Logger LOGGER = LogManager.getLogger("Silent's Gems");
     public static final TextUtil TEXT = new TextUtil(MOD_ID);
-    @Nullable
-    private static CreativeModeTab creativeModeTab;
 
     public SilentGems(IEventBus modEventBus, ModContainer modContainer) {
         Registration.register(modEventBus);
-        modContainer.registerConfig(ModConfig.Type.SERVER, GemsConfig.Common.SPEC);
+        modContainer.registerConfig(ModConfig.Type.COMMON, GemsConfig.COMMON_SPEC);
+
+        if (FMLLoader.getDist() == Dist.CLIENT) {
+            new SilentGemsClient(modContainer);
+        }
     }
 
     public static String getVersion() {