+++ src/main/java/net/silentchaos512/gems/client/SilentGemsClient.java
@@ -1,0 +1,11 @@
+package net.silentchaos512.gems.client;
+
+import net.neoforged.fml.ModContainer;
+import net.neoforged.neoforge.client.gui.ConfigurationScreen;
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+
+public class SilentGemsClient {
+    public SilentGemsClient(ModContainer modContainer) {
+        modContainer.registerExtensionPoint(IConfigScreenFactory.class, ConfigurationScreen::new);
+    }
+}